+++ server/src/main/java/CryptOptima/server/domain/exchange/repository/ExchangeRepository.java
@@ -1,5 +1,6 @@
-package CryptOptima.server.domain.exchange;
+package CryptOptima.server.domain.exchange.repository;
 
+import CryptOptima.server.domain.exchange.entity.Exchange;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
